requirements.txt
aniso8601==9.0.1
asttokens==3.0.0
backcall==0.2.0
blinker==1.8.2
click==8.1.7
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
flask==3.0.3
Flask-RESTful==0.3.9
flask-sqlalchemy==3.0.5
greenlet==3.1.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.4
kafka-python==2.0.2
Mako==1.3.8
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pluggy==1.5.0
prompt-toolkit==3.0.48
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0
pytest==8.3.4
python-dotenv==0.21.0
pytz==2024.2
six==1.17.0
SQLAlchemy==2.0.36
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
typing-extensions==4.12.2
wcwidth==0.2.13
werkzeug==3.0.6
zipp==3.20.2
